requirements.txt

annotated-types==0.7.0
bcrypt==4.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
docopt==0.6.2
email_validator==2.2.0
Flask==3.1.1
Flask-Bcrypt==1.0.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
flask-openapi3==4.2.1
flask-openapi3-elements==9.0.6
flask-openapi3-rapidoc==9.3.8
flask-openapi3-rapipdf==2.2.1
flask-openapi3-redoc==2.5.0
flask-openapi3-scalar==1.34.2
flask-openapi3-swagger==5.27.1
Flask-Reuploaded==1.4.0
Flask-SQLAlchemy==3.1.1
google-api-core==2.25.1
google-api-python-client==2.178.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
greenlet==3.2.4
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.1
MarkupSafe==3.0.2
num2words==0.5.14
oauthlib==3.3.1
passlib==1.7.4
pillow==11.3.0
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
PyMySQL==1.1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.2.0
python-dotenv==1.1.1
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
sib-api-v3-sdk==7.6.0
six==1.17.0
SQLAlchemy==2.0.42
typing-inspection==0.4.1
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.1.3